Factor out ConsentsJourney trait
